+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Supported customized request headers for AI services. (#301)
+- Supported enable automatically extracting full article content from original website. (#306)
 
 ### Changed